keos_project2::eager_pager

Struct EagerPager

pub struct EagerPager {
    mappings: BTreeMap<Va, Mapping>,
}

EagerPager is a struct that implements the Pager trait. It represents a pager strategy that is responsible for eager memory paging.

Fields
mappings: BTreeMap<Va, Mapping>