keos_project2::loader::elf

Struct Phdr

#[repr(C)]
pub struct Phdr {
    pub type_: PType,
    pub p_flags: PFlags,
    pub p_offset: u64,
    pub p_vaddr: u64,
    pub p_paddr: u64,
    pub p_filesz: u64,
    pub p_memsz: u64,
    pub p_align: u64,
}

ELF program header for 64-bit binaries.

Each Phdr entry describes a segment or other information needed for execution.

Fields
type_: PType

Segment type.

p_flags: PFlags

Segment permissions.

p_offset: u64

Offset in the file where the segment starts.

p_vaddr: u64

Virtual address where the segment should be mapped in memory.

p_paddr: u64

Physical address (not commonly used in modern OSes).

p_filesz: u64

Size of the segment in the file.

p_memsz: u64

Size of the segment in memory.

p_align: u64

Alignment of the segment (must be a power of two).

Implementations

impl Phdr

pub fn permission(&self) -> Permission

Get a ELF segment permissions (PFlags) of this Phdr in forms of memory permissions (Permission).

This function translates the permission flags of phdr into the corresponding memory protection flags used by the system and return it. The conversion ensures that the memory is properly set up according to the ELF segmentโs requirements.

Returns

A Permission value representing the phdrโs memory permissions.