keos_project2::page_table

Struct PtIndices

pub struct PtIndices {
    pub va: Va,
    pub pml4ei: usize,
    pub pdptei: usize,
    pub pdei: usize,
    pub ptei: usize,
}

Represents page table indices for a given virtual address (VA).

In the x86_64 architecture, virtual addresses are translated to physical addresses using a 4-level paging hierarchy:

PML4 (Page Map Level 4)

PDPT (Page Directory Pointer Table)

PD (Page Directory)

PT (Page Table)

This structure extracts the index values for each of these levels from a virtual address. This struct provides a way to decompose a virtual address into its respective page table indices.

Fields
va: Va

The virtual address (VA) associated with this page table index breakdown.

pml4ei: usize

Page Map Level 4 Index (PML4EI).

pdptei: usize

Page Directory Pointer table Index (PDPTEI).

pdei: usize

Page Directory Index (PDEI).

ptei: usize

Page Table Index (PTEI).

Implementations

impl PtIndices

pub fn from_va(va: Va) -> Result<Self, PageTableMappingError>

Extracts page table indices from a given virtual address ([Va]).

This function takes a virtual address and calculates the corresponding PML4, PDPT, PD, and PT indices based on x86_64 paging structure.

Arguments

va: A virtual address ([Va]) to be broken down into page table indices.

Returns

Ok(Self): If va is page-aligned (i.e., lower 12 bits are zero).

Err(PageTableMappingError::Unaligned): If va is not page-aligned.