keos_project2::page_table

Struct Walked

pub struct Walked<'a> {
    addr: Va,
    pte: &'a mut Pte,
}

A mutable reference to a page table entry (PTE) associated with a virtual address.

Walked provides safe and convenient access for modifying an existing mapping in the page table. It is typically the result of a successful page table walk and includes both the virtual address and a mutable reference to the corresponding page table entry.

This abstraction is useful for implementing operations such as clearing mappings, updating physical page mappings, or changing permissions.

Fields
addr: Vapte: &'a mut Pte
Implementations

impl Walked<'_>

pub fn clear(&mut self) -> Option<StaleTLBEntry>

Clears the current mapping by returning the physical page and a StaleTLBEntry for flushing the TLB.

Returns

Some(StaleTLBEntry) if the entry is mapped.

None if the entry is not valid.

pub fn set_page( &mut self, page: Page, flags: PteFlags, ) -> Result<(), PageTableMappingError>

Sets this page table entry to map to the given page with the specified flags.

Parameters

page: The physical page to be mapped.

flags: The desired access permissions and attributes for the mapping.

Errors

Returns Err(PageTableMappingError) if the physical address cannot be set (e.g., due to address alignment or capacity limits).