keos_project2::process

Struct Process

pub struct Process {
    pub file_struct: FileStruct,
    pub mm_struct: MmStruct<EagerPager>,
}

A process state of project 2, which contains file struct and mm struct.

Fields
file_struct: FileStructmm_struct: MmStruct<EagerPager>
Implementations

impl Process

pub fn from_mm_struct(mm_struct: MmStruct<EagerPager>) -> Self

Create a process with given MmStruct.

pub fn exit(&self, abi: &SyscallAbi<'_>) -> Result<usize, KernelError>

Exit a process.

This function terminates the calling thread by invoking exit on the current thread. The exit code is provided as the first argument (arg1) of the system call.

Syscall API

int exit(int status);

status: The thread’s exit code.

Parameters

abi: A reference to SyscallAbi, which holds the arguments passed to the system call.

Returns

Never returns.

Behavior

Example

syscall!(SyscallNumber::Exit as usize, 0); // Exit with code 0

Notes

This function does not return in normal execution, as it terminates the thread.

If an error occurs, it returns a KernelError